requirement.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
paypalrestsdk==1.13.3
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
